+++ src/calc.h
@@ -16,10 +16,10 @@ extern "C" {
 
 void process_cmd(char cmd);
 
-void push_decn(const char* signif_str, int16_t exponent, uint8_t no_lift);
+void push_decn(const char* signif_str, int8_t exponent, uint8_t no_lift);
 
 void clear_x(void);
-void set_x(const char* signif_str, int16_t exponent);
+void set_x(const char* signif_str, int8_t exponent);
 __xdata dec80* get_x(void);
 __xdata dec80* get_y(void);
